+++ AltusCaldus/features/environment.py
@@ -1,14 +1,11 @@
 from splinter.browser import Browser
 
 
-
-
 def before_all(context):
     context.browser = Browser('chrome')
     context.server_url = 'http://localhost:8000'
 
 
-
 def after_all(context):
-    context.browser.quit()
+    context.browser.pause()
     context.browser = None
