+++ AltusCaldus/vendas/settings.py
@@ -42,6 +42,7 @@ INSTALLED_APPS = [
     # my apps
     'vendas.servicos',
     'vendas.loja',
+    'behave_django',
 ]
 
 MIDDLEWARE = [
@@ -79,14 +80,13 @@ WSGI_APPLICATION = 'vendas.wsgi.application'
 # Database
 # https://docs.djangoproject.com/en/1.10/ref/settings/#databases
 
-'''
+
 default_dburl = 'sqlite:///' + os.path.join(BASE_DIR, 'db.sqlite3')
 DATABASES = {
     'default': config('DATABASE_URL', default=default_dburl, cast=dburl),
 }
 
 '''
-
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
@@ -97,7 +97,7 @@ DATABASES = {
         'PORT': '5432',
     }
 }
-
+'''
 
 
 # Password validation
